Change icons to Lucide + start general table
This commit changes the icon pack from Feather to Lucide as it is under
more development with new icons (like for sorting) that were missing. As
a successor to Feather, it has a very similar feel.

Also started work on a table that can be extended to fit more general
purposes to replace the already-existing tables. This one will have
sorting capabilities and "copy to clipboard".